--- code/__init__.py
@@ -1,3 +1,0 @@
-import os
-import sys
-sys.path.insert(0, os.getcwd() + "/code")
