+++ requirements.txt
@@ -8,3 +8,4 @@ pytest-mock
 python-telegram-bot
 python-telegram-bot-calendar
 currencyapicom
+pymongo
