+++ telebot_code/__init__.py
@@ -1,0 +1,3 @@
+import os
+import sys
+sys.path.insert(0, os.getcwd() + "/telebot_code")
