+++ test/test_display.py
@@ -2,7 +2,7 @@ import os
 import json
 from unittest.mock import patch
 from telebot import types
-from code import display
+from telebot_code import display
 
 
 @patch('telebot.telebot')
